完善

+++ OznerLibrary/Wifi/MXChip/libmosquitto/send_client_mosq.c
@@ -57,7 +57,7 @@ int _mosquitto_send_connect(struct mosquitto *mosq, uint16_t keepalive, bool cle
 	packet = _mosquitto_calloc(1, sizeof(struct _mosquitto_packet));
 	if(!packet) return MOSQ_ERR_NOMEM;
 
-	payloadlen = 2+strlen(mosq->id);
+	payloadlen = 2+(int)strlen(mosq->id);
 	if(mosq->will){
 		will = 1;
 		assert(mosq->will->topic);
@@ -152,7 +152,7 @@ int _mosquitto_send_subscribe(struct mosquitto *mosq, int *mid, bool dup, const 
 	packet = _mosquitto_calloc(1, sizeof(struct _mosquitto_packet));
 	if(!packet) return MOSQ_ERR_NOMEM;
 
-	packetlen = 2 + 2+strlen(topic) + 1;
+	packetlen = 2 + 2+(uint32_t)strlen(topic) + 1;
 
 	packet->command = SUBSCRIBE | (dup<<3) | (1<<1);
 	packet->remaining_length = packetlen;
@@ -197,7 +197,7 @@ int _mosquitto_send_unsubscribe(struct mosquitto *mosq, int *mid, bool dup, cons
 	packet = _mosquitto_calloc(1, sizeof(struct _mosquitto_packet));
 	if(!packet) return MOSQ_ERR_NOMEM;
 
-	packetlen = 2 + 2+strlen(topic);
+	packetlen = 2 + 2+(uint32_t)strlen(topic);
 
 	packet->command = UNSUBSCRIBE | (dup<<3) | (1<<1);
 	packet->remaining_length = packetlen;
