+++ OznerLibrary/Wifi/MXChip/libmosquitto/send_mosq.c
@@ -260,7 +260,7 @@ int _mosquitto_send_real_publish(struct mosquitto *mosq, uint16_t mid, const cha
 	assert(mosq);
 	assert(topic);
 
-	packetlen = 2+strlen(topic) + payloadlen;
+	packetlen = 2+(int)strlen(topic) + payloadlen;
 	if(qos > 0) packetlen += 2; /* For message id */
 	packet = _mosquitto_calloc(1, sizeof(struct _mosquitto_packet));
 	if(!packet) return MOSQ_ERR_NOMEM;
